Stats update + Recovery and reset command added + new Riot api system implemented + Bug Fix

+++ src/main/java/ch/kalunight/zoe/service/InfoPanelRefresher.java
@@ -1,15 +1,15 @@
 package ch.kalunight.zoe.service;
 
+import java.time.OffsetDateTime;
 import java.util.ArrayList;
 import java.util.List;
-
+import java.util.stream.Collectors;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
 import com.jagrosh.jdautilities.command.CommandEvent;
-
 import ch.kalunight.zoe.ServerData;
 import ch.kalunight.zoe.Zoe;
+import ch.kalunight.zoe.model.ControlPannel;
 import ch.kalunight.zoe.model.InfoCard;
 import ch.kalunight.zoe.model.Player;
 import ch.kalunight.zoe.model.Server;
@@ -43,41 +43,81 @@ public class InfoPanelRefresher implements Runnable {
     try {
       if(server.getInfoChannel() != null) {
         if(server.getControlePannel().getInfoPanel().isEmpty()) {
-          server.getControlePannel().getInfoPanel().add(server.getInfoChannel().sendMessage("__**Control Panel**__\n \n*Loading...*").complete());
+          server.getControlePannel().getInfoPanel()
+          .add(server.getInfoChannel().sendMessage("__**Control Panel**__\n \n*Loading...*").complete());
         }
 
         ArrayList<String> infoPanels = CommandEvent.splitMessage(refreshPannel());
 
-        if(infoPanels.size() < server.getControlePannel().getInfoPanel().size()) {
-          int nbrMessageToDelete = server.getControlePannel().getInfoPanel().size() - infoPanels.size();
-          for(int i = 0; i < nbrMessageToDelete; i++) {
-            Message message = server.getControlePannel().getInfoPanel().get(i);
-            message.delete().queue();
-            server.getControlePannel().getInfoPanel().remove(message);
+        if(server.getInfoChannel() != null || server.getGuild().getTextChannelById(server.getInfoChannel().getId()) != null) {
+
+          if(infoPanels.size() < server.getControlePannel().getInfoPanel().size()) {
+            int nbrMessageToDelete = server.getControlePannel().getInfoPanel().size() - infoPanels.size();
+            for(int i = 0; i < nbrMessageToDelete; i++) {
+              Message message = server.getControlePannel().getInfoPanel().get(i);
+              message.delete().queue();
+              server.getControlePannel().getInfoPanel().remove(message);
+            }
+
+          } else {
+            int nbrMessageToAdd = infoPanels.size() - server.getControlePannel().getInfoPanel().size();
+            for(int i = 0; i < nbrMessageToAdd; i++) {
+              server.getControlePannel().getInfoPanel().add(server.getInfoChannel().sendMessage("loading ...").complete());
+            }
           }
 
-        }else {
-          int nbrMessageToAdd = infoPanels.size() - server.getControlePannel().getInfoPanel().size();
-          for(int i = 0; i < nbrMessageToAdd; i++) {
-            server.getControlePannel().getInfoPanel().add(server.getInfoChannel().sendMessage("loading ...").complete());
+          for(int i = 0; i < infoPanels.size(); i++) {
+            server.getControlePannel().getInfoPanel().get(i).editMessage(infoPanels.get(i)).queue();
           }
-        }
 
-        for(int i = 0; i < infoPanels.size(); i++) {
-          server.getControlePannel().getInfoPanel().get(i).editMessage(infoPanels.get(i)).queue();
-        }
+          manageInfoCards();
 
-        manageInfoCards();
+          cleaningInfoChannel();
+        } else {
+          server.setInfoChannel(null);
+          server.setControlePannel(new ControlPannel());
+        }
       }
-    }catch (NullPointerException e){
-      logger.info("The Thread has crashed normally because of deletion of infoChannel : {}", e);
-    }catch (Exception e) {
-      logger.warn("The thread got a unexpected error : {}", e);
-    }finally {
+    } catch(NullPointerException e) {
+      logger.info("The Thread has crashed normally because of deletion of infoChannel :", e);
+    } catch(Exception e) {
+      logger.warn("The thread got a unexpected error :", e);
+    } finally {
       ServerData.getServersIsInTreatment().put(server.getGuild().getId(), false);
     }
   }
 
+
+  private void cleaningInfoChannel() {
+
+    List<Message> messagesToCheck = server.getInfoChannel().getIterableHistory().stream()
+        .limit(1000)
+        .filter(m-> m.getAuthor().getId().equals(Zoe.getJda().getSelfUser().getId()))
+        .collect(Collectors.toList());
+
+    List<Message> messagesToDelete = new ArrayList<>();
+    
+    for(Message messageToCheck : messagesToCheck) {
+      if(!messageToCheck.getCreationTime().isBefore(OffsetDateTime.now().minusHours(1)) || server.getControlePannel().getInfoPanel().contains(messageToCheck)) {
+        continue;
+      }
+      
+      messagesToDelete.add(messageToCheck);
+    }
+    
+    if(messagesToDelete.isEmpty()) {
+      return;
+    }
+    
+    if(messagesToDelete.size() > 1) {
+      server.getInfoChannel().purgeMessages(messagesToDelete);
+    }else {
+      for(Message messageToDelete : messagesToDelete) {
+        messageToDelete.delete().queue();
+      }
+    }
+    
+  }
 
   private void manageInfoCards() {
     TextChannel controlPannel = server.getInfoChannel();
@@ -104,10 +144,10 @@ public class InfoPanelRefresher implements Runnable {
       try {
         try {
           cardsToRemove.get(i).getMessage().delete().complete();
-        }catch(ErrorResponseException  e) {
+        } catch(ErrorResponseException e) {
           if(e.getErrorResponse() == ErrorResponse.UNKNOWN_MESSAGE) {
             logger.info("Message already deleted : {}", e.getMessage());
-          }else {
+          } else {
             logger.warn("Unhandle error : {}", e.getMessage());
             throw e;
           }
@@ -115,17 +155,17 @@ public class InfoPanelRefresher implements Runnable {
 
         try {
           cardsToRemove.get(i).getTitle().delete().complete();
-        }catch(ErrorResponseException  e) {
+        } catch(ErrorResponseException e) {
           if(e.getErrorResponse() == ErrorResponse.UNKNOWN_MESSAGE) {
             logger.info("Message already deleted : {}", e.getMessage());
-          }else {
+          } else {
             logger.warn("Unhandle error : {}", e.getMessage());
             throw e;
           }
         }
         server.getControlePannel().getInfoCards().remove(cardsToRemove.get(i));
       } catch(Exception e) {
-        logger.warn("Impossible de delete message, retry next time : {}", e.getMessage(), e);
+        logger.warn("Impossible to delete message, retry next time : {}", e.getMessage(), e);
       }
     }
   }
@@ -147,13 +187,14 @@ public class InfoPanelRefresher implements Runnable {
           InfoCard card = null;
 
           if(listOfPlayerInTheGame.size() == 1) {
-            MessageEmbed messageCard =
-                MessageBuilderRequest.createInfoCard1summoner(player.getDiscordUser(), player.getSummoner(), currentGameInfo, player.getRegion());
+            MessageEmbed messageCard = MessageBuilderRequest.createInfoCard1summoner(player.getDiscordUser(), player.getSummoner(),
+                currentGameInfo, player.getRegion());
             if(messageCard != null) {
               card = new InfoCard(listOfPlayerInTheGame, messageCard);
             }
           } else if(listOfPlayerInTheGame.size() > 1) {
-            MessageEmbed messageCard = MessageBuilderRequest.createInfoCardsMultipleSummoner(listOfPlayerInTheGame, currentGameInfo, player.getRegion());
+            MessageEmbed messageCard =
+                MessageBuilderRequest.createInfoCardsMultipleSummoner(listOfPlayerInTheGame, currentGameInfo, player.getRegion());
 
             if(messageCard != null) {
               card = new InfoCard(listOfPlayerInTheGame, messageCard);
@@ -234,8 +275,7 @@ public class InfoPanelRefresher implements Runnable {
       List<Player> playersList = team.getPlayers();
 
       for(Player player : playersList) {
-        stringMessage
-        .append(player.getSummoner().getName() + " (" + player.getDiscordUser().getAsMention() + ") : ");
+        stringMessage.append(player.getSummoner().getName() + " (" + player.getDiscordUser().getAsMention() + ") : ");
 
         CurrentGameInfo actualGame = server.getCurrentGames().get(player.getSummoner().getId());
